Changed all double's to float's

+++ src/particles/GroundParticle.cc
@@ -1,4 +1,4 @@
-//  $Id: GroundParticle.cc,v 1.2 2000/02/09 21:43:44 grumbel Exp $
+//  $Id: GroundParticle.cc,v 1.3 2000/03/08 01:37:22 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 1999 Ingo Ruhnke <grumbel@gmx.de>
@@ -21,7 +21,7 @@
 #include "../PingusResource.hh"
 #include "GroundParticle.hh"
 
-GroundParticle::GroundParticle(int x, int y, double x_a, double y_a)
+GroundParticle::GroundParticle(int x, int y, float x_a, float y_a)
 {
   surface = CL_Surface::load("Particles/ground", PingusResource::get("pingus.dat"));
   x_pos = x;
@@ -44,7 +44,7 @@ GroundParticle::let_move(void)
 }
 
 void
-GroundParticle::draw_offset(int ofx, int ofy, double s) const
+GroundParticle::draw_offset(int ofx, int ofy, float s) const
 {
   if (fast_mode)
     return;
