+++ src/particles/ParticleHolder.hh
@@ -1,4 +1,4 @@
-//  $Id: ParticleHolder.hh,v 1.3 2000/02/11 16:58:28 grumbel Exp $
+//  $Id: ParticleHolder.hh,v 1.4 2000/03/08 01:37:22 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 1999 Ingo Ruhnke <grumbel@gmx.de>
@@ -35,13 +35,13 @@ public:
   virtual ~ParticleHolder();
 
   void let_move(void);
-  void draw_offset(int, int, double); // const;
+  void draw_offset(int, int, float); // const;
   void add_pingu_explo(int x, int y);
   void add_particle(Particle* p);
   
   void init_particles();
   void clear_particles();
-  void add_pingu_particle(int, int, double, double);
+  void add_pingu_particle(int, int, float, float);
 };
 
 #endif
